requirements.txt
mido
numpy
pytorch
torchaudio
torchvision
soundfile
pretty_midi